+++ packages/babel-plugin-s2s-initial-state-creater/src/index.js
@@ -3,7 +3,7 @@ import { parseExpression } from 'babylon'
 import { inheritsOpts } from 's2s-utils'
 import { flowFakerSync } from 'flow-faker'
 import generator from 'babel-generator'
-import type { BabelPath, State } from 'types'
+import type { BabelPath, State } from 'types/babel'
 import json5 from 'json5'
 import typedAssign from 'typed-assign'
 // import blog from 'babel-log'
